feat: extracted utility classes out of opgg.py
- moved all static methods out of opgg.py and into utils.py
