+++ opgg/opgg.py
@@ -10,15 +10,14 @@ import logging
 import requests
 import traceback
 from datetime import datetime
-from bs4 import BeautifulSoup
 from typing import Literal
 
 from opgg.summoner import Summoner
 from opgg.season import Season, SeasonInfo
-from opgg.champion import ChampionStats, Champion, Spell, Passive, Skin, Price
+from opgg.champion import ChampionStats, Champion
 from opgg.league_stats import LeagueStats, Tier, QueueInfo
 from opgg.game import Game
-from opgg.params import Region, By
+from opgg.params import Region
 from opgg.cacher import Cacher
 
 
@@ -27,7 +26,7 @@ class OPGG:
     ### OPGG.py
     A simple library to access and structure the data from OP.GG's Website & API.
     
-    Copyright (c) 2023, ShoobyDoo
+    Copyright (c) 2023-2024, ShoobyDoo
     License: BSD-3-Clause, See LICENSE for more details.
     """
     
@@ -196,7 +195,7 @@ class OPGG:
         self.logger.debug(f"self._games_api_url = {self._games_api_url}")
     
     
-    def get_summoner(self) -> Summoner:
+    def get_summoner(self, return_content_only = False) -> Summoner:
         """
         A method to get data from the OPGG API and form a Summoner object.
         
@@ -207,7 +206,7 @@ class OPGG:
             
         ### Returns:
             `Summoner`: A Summoner object representing the summoner.
-        """
+        """        
         self.logger.info(f"Sending request to OPGG API... (API_URL = {self.api_url}, HEADERS = {self.headers})")
         res = requests.get(self.api_url, headers=self.headers)
         
@@ -221,6 +220,11 @@ class OPGG:
             content = json.loads(res.text)["data"]
         else:
             res.raise_for_status()
+        
+        # If return_res is passed in func args, return the content
+        # Required in tests to get all the raw content without building the summoner object
+        if (return_content_only):
+            return content
         
         try:            
             for season in content["summoner"]["previous_seasons"]:
@@ -300,32 +304,6 @@ class OPGG:
                     penta_kill = champion["penta_kill"],
                     game_length_second = champion["game_length_second"]
                 ))
-            
-            # if (len(content["recent_game_stats"]) > 0):
-            #     for game in content["recent_game_stats"]:
-            #         tmp_champ = None
-            #         if self.all_champions:
-            #             for _champion in self.all_champions:
-            #                 if _champion.id == game["champion_id"]:
-            #                     tmp_champ = _champion
-            #                     break
-                        
-            #         recent_game_stats.append(Game(
-            #             game_id = game["game_id"],
-            #             champion = tmp_champ,
-            #             kill = game["kill"],
-            #             death = game["death"],
-            #             assist = game["assist"],
-            #             position = game["position"],
-            #             is_win = game["is_win"],
-            #             is_remake = game["is_remake"],
-            #             op_score = game["op_score"],
-            #             op_score_rank = game["op_score_rank"],
-            #             is_opscore_max_in_team = game["is_opscore_max_in_team"],
-            #             created_at = game["created_at"]
-            #         ))
-                    
-            # else:
             
             # page props did not return any recent games, lets query the /games endpoint instead
             # gets the summoner id from the objects internal self._game_api_url's self.summoner_id ref
@@ -471,7 +449,7 @@ class OPGG:
         return summoners if len(summoners) > 1 else summoners[0]
 
     
-    def get_recent_games(self, results: int = 10, game_type: Literal["total", "ranked", "normal"] = "total") -> list[Game]:
+    def get_recent_games(self, results: int = 10, game_type: Literal["total", "ranked", "normal"] = "total", return_content_only = False) -> list[Game]:
         recent_games = []
         res = requests.get(f"{self._games_api_url}?&limit={results}&game_type={game_type}", headers=self.headers)
         
@@ -481,10 +459,13 @@ class OPGG:
         else:
             res.raise_for_status()
         
+        if return_content_only:
+            return game_data
+        
         try:
             for game in game_data:
                 tmp_game = Game(
-                    game_id = game["id"],
+                    id = game["id"],
                     champion = next((champion for champion in self.all_champions if champion.id == game["myData"]["champion_id"]), None),
                     kill = int(game["myData"]["stats"]["kill"]),
                     death = int(game["myData"]["stats"]["death"]),
@@ -506,268 +487,4 @@ class OPGG:
             pass
     
 
-    @staticmethod
-    def get_page_props(summoner_names: str | list[str] = "", region = Region.NA) -> dict:
-        """
-        Get the page props from OPGG. (Contains data such as summoner info, champions, seasons, etc.)
-        
-        ### Parameters
-            summoner_names : `str | list[str], optional`
-                Pass a single or comma separated `str` or a list of summoner names.\n
-                Note: Default is "abc", as this can be any valid summoner if you just want page props. (All champs, seasons, etc.)
-            
-            region : `Region, optional`
-                Pass the region you want to search in. Default is "NA".
-
-        ### Returns
-            `dict` : Returns a dictionary with the page props.
-        """
-        
-        if isinstance(summoner_names, list): summoner_names = ",".join(summoner_names)
-        
-        url = f"https://op.gg/multisearch/{region}?summoners={summoner_names}"
-        headers = { 
-            "User-Agent": "Mozilla/5.0 (Windows NT 10.1; en-US) AppleWebKit/602.14 (KHTML, like Gecko) Chrome/55.0.2575.244 Safari/603.4 Edge/11.82011",
-        }
-
-        res = requests.get(url, headers=headers, allow_redirects=True)
-        soup = BeautifulSoup(res.content, "html.parser")
-        
-        return json.loads(soup.select_one("#__NEXT_DATA__").text)['props']['pageProps']
     
-    
-    @staticmethod
-    def get_all_seasons(region = Region.NA, page_props: dict = None) -> list[SeasonInfo]:
-        """
-        Get all seasons from OPGG.
-
-        ### Args:
-            region : `Region, optional`
-                Pass the region you want to search in. Defaults to "NA".
-                
-            page_props : `dict, optional`
-                Pass the page props if the program has queried them once before.\n
-                Note: Defaults to None, but if you pass them it reduces the overhead of another request out to OPGG.
-
-        ### Returns:
-            `list[SeasonInfo]` : A list of SeasonInfo objects.
-        """
-        # Check cache first, and if found return that instead of querying.
-        cached_seasons = Cacher().get_all_seasons()
-        if cached_seasons:
-            return cached_seasons
-        
-        if page_props == None and not OPGG.cached_page_props:
-            page_props = OPGG.get_page_props(region)
-            OPGG.cached_page_props = page_props
-            
-        elif OPGG.cached_page_props:
-            page_props = OPGG.cached_page_props
-        
-        seasons = []
-        for season in dict(page_props['seasonsById']).values():
-            seasons.append(SeasonInfo(
-                id = season["id"],
-                value = season["value"],
-                display_value = season["display_value"],
-                is_preseason = season["is_preseason"]
-            ))
-        
-        return seasons
-
-
-    @staticmethod
-    def get_season_by(by: By, value: int | str | list) -> SeasonInfo | list[SeasonInfo]:
-        """
-        Get a season by a specific metric.
-        
-        ### Args:
-            by : `By`
-                Pass a By enum to specify how you want to get the season(s).
-            
-            value : `int | str | list`
-                Pass the value(s) you want to search by. (id, display_value, etc.)
-        
-        ### Returns:
-            `SeasonInfo | list[SeasonInfo]` : A single or list of SeasonInfo objects.
-        """
-        
-        all_seasons = OPGG.get_all_seasons()
-        result_set = []
-        
-        if by == By.ID:
-            if isinstance(value, list):
-                for season in all_seasons:
-                    for id in value:
-                        if season.id == id:
-                            result_set.append(season)
-            else:
-                for season in all_seasons:
-                    if season.id == int(value):
-                        result_set.append(season)
-        
-        # TODO: perhaps add more ways to get season objs, like by is_preseason, or display_name, etc.           
-        
-        return result_set if len(result_set) > 1 else result_set[0]
-    
-    
-    @staticmethod
-    def get_all_champions(region = Region.NA, page_props: dict = None) -> list[Champion]:
-        """
-        Get all champion info from OPGG.\n
-        Page props method will be deprecated very soon in favour of simply pinging a champion endpoint I found.
-
-        ### Args:
-            region : `Region, optional`
-                Pass the region you want to search in. Defaults to "NA".
-                
-            page_props : `dict, optional`
-                Pass the page props if the program has queried them once before.\n
-                Note: Defaults to None, but if you pass them it reduces the overhead of another request out to OPGG.
-
-        Returns:
-            `list[Champion]` : A list of Champion objects.
-        """
-        # Check cache, if found, return it, otherwise continue to below logic.
-        cached_champions = Cacher().get_all_champs()
-        if cached_champions:
-            return cached_champions
-        
-        if page_props == None and not OPGG.cached_page_props:
-            # pass any valid username here, it doesnt matter.
-            # we just need the page props, and we dont get them without an actual user
-            page_props = OPGG.get_page_props(region)
-            OPGG.cached_page_props = page_props
-            
-        elif OPGG.cached_page_props:
-            page_props = OPGG.cached_page_props
-            
-        champions = []
-        
-        for champion in dict(page_props["championsById"]).values():
-            # reset per iteration
-            _spells = []
-            _skins = []
-            
-            for skin in champion["skins"]:
-                _prices = []
-                
-                if skin["prices"]:
-                    for price in skin["prices"]:
-                        _prices.append(Price(
-                            currency = price["currency"] if "RP" in price["currency"] else "BE",
-                            cost = price["cost"]
-                        ))
-                else:
-                    _prices = None
-                
-                _skins.append(Skin(
-                    id = skin["id"],
-                    name = skin["name"],
-                    centered_image = skin["centered_image"],
-                    skin_video_url = skin["skin_video_url"],
-                    prices = _prices,
-                    sales = skin["sales"]
-                ))
-                    
-            for spell in champion["spells"]:
-                _spells.append(Spell(
-                    key = spell["key"],
-                    name = spell["name"],
-                    description = spell["description"],
-                    max_rank = spell["max_rank"],
-                    range_burn = spell["range_burn"],
-                    cooldown_burn = spell["cooldown_burn"],
-                    cost_burn = spell["cost_burn"],
-                    tooltip = spell["tooltip"],
-                    image_url = spell["image_url"],
-                    video_url = spell["video_url"]
-                ))       
-              
-            champions.append(Champion(
-                id = champion["id"],
-                key = champion["key"],
-                name = champion["name"],
-                image_url = champion["image_url"],
-                evolve = champion["evolve"],
-                passive = Passive(
-                    name = champion["passive"]["name"],
-                    description = champion["passive"]["description"],
-                    image_url = champion["passive"]["image_url"],
-                    video_url = champion["passive"]["video_url"]
-                ),
-                spells = _spells,
-                skins = _skins
-            ))            
-        
-        return champions
-    
-    
-    @staticmethod
-    def get_champion_by(by: By, value: int | str | list, **kwargs) -> Champion | list[Champion]:
-        """
-        Get a single or list of champions by a specific metric.
-        
-        ### Args:
-            by : `By`
-                Pass a By enum to specify how you want to get the champion(s).
-                
-            value : `int | str | list`
-                Pass the value(s) you want to search by. (id, key, name, etc.)
-                
-            **kwargs : `any`
-                Pass any additional keyword arguments to narrow down the search.\n
-                Note: Currently only supports "currency" for the cost of a champion.\n
-                
-                Example: 
-                    `get_champion_by(By.COST, 450, currency=By.BLUE_ESSENCE)`
-        """
-        # Currently kwargs only handles "currency" for the cost of a champion,
-        # but I might introduce other metrics of getting champ objs later, idk...
-        
-        all_champs = OPGG.get_all_champions()
-        result_set = []
-        
-        if by == By.ID:
-            if isinstance(value, list):
-                for champ in all_champs:
-                    for id in value:
-                        if champ.id == id:
-                            result_set.append(champ)
-            else:
-                for champ in all_champs:
-                    if champ.id == int(value):
-                        result_set.append(champ)
-                
-        elif by == By.KEY:
-            if isinstance(value, list):
-                for champ in all_champs:
-                    for key in value:
-                        if champ.key == key:
-                            result_set.append(champ)
-            else:
-                for champ in all_champs:
-                    if champ.key == value:
-                        result_set.append(champ)
-        
-        elif by == By.NAME:
-            if isinstance(value, list):
-                for champ in all_champs:
-                    for name in value:
-                        if champ.name == name:
-                            result_set.append(champ)
-            else:
-                for champ in all_champs:
-                    if champ.name == value:
-                        result_set.append(champ)
-        
-        elif by == By.COST:
-            for champ in all_champs:
-                if champ.skins[0].prices:
-                    for price in champ.skins[0].prices:
-                        if str(kwargs["currency"]).upper() == price.currency and price.cost in value:
-                            result_set.append(champ)
-                
-        
-        # if the result set is larger than one, return the whole list, otherwise just return the object itself.
-        return result_set if len(result_set) > 1 else result_set[0]
